Quote links, move things around, tweaks
This moves some Auth0 stuff around so more of it is in the auth0 app. Further,
it fixes auth0 url generation so that the values are quoted and it fixes a
typo (redirect_url -> redirect_uri).
