+++ standup/status/views.py
@@ -21,7 +21,6 @@ from django.views.generic import DetailView, TemplateView, UpdateView
 
 from raven.contrib.django.models import client
 
-from standup.auth0.utils import is_auth0_configured
 from standup.status.forms import StatusizeForm, ProfileForm
 from standup.status.models import Status, Team, Project, StandupUser
 from standup.status.utils import enddate, startdate
@@ -158,19 +157,6 @@ class ProfileView(UpdateView):
 
 class LoginView(TemplateView):
     template_name = 'users/login.html'
-
-    def get_context_data(self, **kwargs):
-        ctx = super().get_context_data(**kwargs)
-        ctx['auth0_configured'] = is_auth0_configured()
-
-        ctx['auth0_login_url'] = settings.AUTH0_LOGIN_URL.format(
-            AUTH0_DOMAIN=settings.AUTH0_DOMAIN,
-            AUTH0_CLIENT_ID=settings.AUTH0_CLIENT_ID,
-            AUTH0_CALLBACK_URL=settings.AUTH0_CALLBACK_URL,
-            # FIXME(willkg): This should be a token that ties both ends.
-            STATE='foo',
-        )
-        return ctx
 
     def get(self, request, *args, **kwargs):
         if request.user.is_authenticated():
